ISSUE #3307 - rename placeholder cards
